+++ tests/test_errors.py
@@ -1,0 +1,6 @@
+from pyfomod.errors import ErrorKind
+
+
+def test_compare():
+    assert ErrorKind.NOTE < ErrorKind.WARNING
+    assert ErrorKind.WARNING < ErrorKind.ERROR