+++ tests/test_parser.py
@@ -1,7 +1,8 @@
 import textwrap
 from pathlib import Path
 
-from pyfomod import ValidationWarning, parser
+from pyfomod import parser
+from pyfomod.errors import ErrorID, ErrorKind
 
 PACKAGE_PATH = Path(__file__).parent / "package_test"
 INFO_PATH = Path(PACKAGE_PATH, "fomod", "info.xml")
@@ -25,11 +26,14 @@ def test_preserve_data(tmp_path):
     assert orig_conf == new_conf
 
 
-def test_parse(tmp_path):
+def test_tuple_parse():
     root = parser.parse(str(PACKAGE_PATH))
     tuple_root = parser.parse((str(INFO_PATH), str(CONF_PATH)))
     assert root.to_string() == tuple_root.to_string()
     assert root._info.to_string() == tuple_root._info.to_string()
+
+
+def test_parse(tmp_path):
     content = textwrap.dedent(
         """\
             <config>
@@ -65,140 +69,164 @@ def test_parse(tmp_path):
     with conf_path.open("w") as conf_file:
         conf_file.write(content)
         conf_file.write("\n")
-    warnings = []
-    root = parser.parse((None, str(conf_path)), warnings=warnings)
+    errors = []
+    root = parser.parse((None, str(conf_path)), errors=errors)
     expected = [
-        ValidationWarning(
-            "XML Syntax Error",
-            "Element 'moduleDependencies': This element is not expected. "
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.INVALID_SYNTAX,
+            "title": "XML Syntax Error",
+            "msg": "Element 'moduleDependencies': This element is not expected. "
             "Expected is ( moduleName ).",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "XML Comments Present",
-            "There are comments in the fomod, they will be ignored.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Invalid Condition Type",
-            "Condition Type was set to 'invalid' in tag 'moduleDependencies' "
+        },
+        {
+            "kind": ErrorKind.WARNING,
+            "elem": None,
+            "id": ErrorID.COMMENTS_PRESENT,
+            "title": "XML Comments Present",
+            "msg": "There are comments in the fomod, they will be ignored.",
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.conditions,
+            "id": ErrorID.INVALID_ENUM,
+            "title": "Invalid Condition Type Attribute",
+            "msg": "Condition Type was set to 'invalid' in tag 'moduleDependencies' "
             "but can only be one of: 'And', 'Or'. "
             "It was set to default 'And'.",
-            root.conditions,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing File Attribute",
-            "The 'file' attribute on the 'fileDependency' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.REQUIRED_ATTRIBUTE,
+            "title": "Missing File Attribute",
+            "msg": "The 'file' attribute on the 'fileDependency' "
             "tag is required. This tag will be skipped.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing State Attribute",
-            "The 'state' attribute on the 'fileDependency' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.DEFAULT_ATTRIBUTE,
+            "title": "Missing State Attribute",
+            "msg": "The 'state' attribute on the 'fileDependency' "
             "tag is required. It was set to 'Active'.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Invalid File Type",
-            "File Type was set to 'invalid' in tag 'fileDependency' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.INVALID_ENUM,
+            "title": "Invalid File Type Attribute",
+            "msg": "File Type was set to 'invalid' in tag 'fileDependency' "
             "but can only be one of: 'Active', 'Inactive', 'Missing'. "
             "It was set to default 'Active'.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Flag Attribute",
-            "The 'flag' attribute on the 'flagDependency' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.REQUIRED_ATTRIBUTE,
+            "title": "Missing Flag Attribute",
+            "msg": "The 'flag' attribute on the 'flagDependency' "
             "tag is required. This tag will be skipped.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Value Attribute",
-            "The 'value' attribute on the 'flagDependency' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.DEFAULT_ATTRIBUTE,
+            "title": "Missing Value Attribute",
+            "msg": "The 'value' attribute on the 'flagDependency' "
             "tag is required. It was set to ''.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Version Attribute",
-            "The 'version' attribute on the 'gameDependency' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.REQUIRED_ATTRIBUTE,
+            "title": "Missing Version Attribute",
+            "msg": "The 'version' attribute on the 'gameDependency' "
             "tag is required. This tag will be skipped.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Source Attribute",
-            "The 'source' attribute on the 'file' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.REQUIRED_ATTRIBUTE,
+            "title": "Missing Source Attribute",
+            "msg": "The 'source' attribute on the 'file' "
             "tag is required. This tag will be skipped.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Invalid Order",
-            "Order was set to 'invalid' in tag 'installSteps' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.pages,
+            "id": ErrorID.INVALID_ENUM,
+            "title": "Invalid Order Attribute",
+            "msg": "Order was set to 'invalid' in tag 'installSteps' "
             "but can only be one of: "
             "'Ascending', 'Descending', 'Explicit'. "
             "It was set to default 'Ascending'.",
-            root.pages,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Name Attribute",
-            "The 'name' attribute on the 'installStep' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.pages[0],
+            "id": ErrorID.DEFAULT_ATTRIBUTE,
+            "title": "Missing Name Attribute",
+            "msg": "The 'name' attribute on the 'installStep' "
             "tag is required. It was set to ''.",
-            root.pages[0],
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Name Attribute",
-            "The 'name' attribute on the 'group' tag is required. It was set to ''.",
-            root.pages[0][0],
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Type Attribute",
-            "The 'type' attribute on the 'group' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.pages[0][0],
+            "id": ErrorID.DEFAULT_ATTRIBUTE,
+            "title": "Missing Name Attribute",
+            "msg": "The 'name' attribute on the 'group' tag "
+            "is required. It was set to ''.",
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.pages[0][0],
+            "id": ErrorID.DEFAULT_ATTRIBUTE,
+            "title": "Missing Type Attribute",
+            "msg": "The 'type' attribute on the 'group' "
             "tag is required. It was set to 'SelectAny'.",
-            root.pages[0][0],
-            critical=True,
-        ),
-        ValidationWarning(
-            "Invalid Group Type",
-            "Group Type was set to 'invalid' in tag 'group' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.pages[0][1],
+            "id": ErrorID.INVALID_ENUM,
+            "title": "Invalid Group Type Attribute",
+            "msg": "Group Type was set to 'invalid' in tag 'group' "
             "but can only be one of: "
             "'SelectAny', 'SelectAll', 'SelectAtLeastOne', "
             "'SelectAtMostOne', 'SelectExactlyOne'. "
             "It was set to default 'SelectAny'.",
-            root.pages[0][1],
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Name Attribute",
-            "The 'name' attribute on the 'plugin' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": root.pages[0][1][0],
+            "id": ErrorID.DEFAULT_ATTRIBUTE,
+            "title": "Missing Name Attribute",
+            "msg": "The 'name' attribute on the 'plugin' "
             "tag is required. It was set to ''.",
-            root.pages[0][1][0],
-            critical=True,
-        ),
-        ValidationWarning(
-            "Invalid Order",
-            "Order was set to 'invalid' in tag 'optionalFileGroups' "
+        },
+        {
+            "kind": ErrorKind.ERROR,
+            "elem": None,
+            "id": ErrorID.INVALID_ENUM,
+            "title": "Invalid Order Attribute",
+            "msg": "Order was set to 'invalid' in tag 'optionalFileGroups' "
             "but can only be one of: "
             "'Ascending', 'Descending', 'Explicit'. "
             "It was set to default 'Ascending'.",
-            None,
-            critical=True,
-        ),
-        ValidationWarning(
-            "Missing Info XML",
-            "Info.xml is missing from the fomod subfolder.",
-            None,
-            critical=False,
-        ),
+        },
+        {
+            "kind": ErrorKind.NOTE,
+            "elem": None,
+            "id": ErrorID.MISSING_INFO,
+            "title": "Missing Info XML",
+            "msg": "Info.xml is missing from the fomod subfolder.",
+        },
     ]
-    assert warnings == expected
+    for expect, actual in zip(expected, errors):
+        assert expect["kind"] is actual.kind
+        assert expect["elem"] is actual.elem
+        assert expect["id"] is actual.id
+        assert expect["title"] == actual.title
+        assert expect["msg"] == actual.msg